+++ src/request_hash.js
@@ -2,9 +2,10 @@ var crypto = require('crypto');
 var Util = require('./util');
 var HeaderUtil = require('./header_util');
 
-var RequestHash = function (request, cacheHeaders) {
+var RequestHash = function (request, cacheHeaders, whiteLabel) {
   this.request = request;
   this.cacheHeaders = cacheHeaders;
+  this.whiteLabel = whiteLabel;
 };
 
 
@@ -19,6 +20,10 @@ RequestHash.prototype.toString = function () {
 RequestHash.prototype._filteredAttributes = function () {
   var filteredAttributes = Util.simpleCopy(this.request);
   filteredAttributes.headers = HeaderUtil.filterHeaders(this.cacheHeaders, this.request.headers);
+  if (this.whiteLabel) {
+    filteredAttributes.hostname = 'example.com';
+    filteredAttributes.port = 80;
+  }
   return filteredAttributes;
 }